// web/projects/portal/src/app/graph/model/chart-object.ts
/*
 * This file is part of StyleBI.
 * Copyright (C) 2024  InetSoft Technology
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
import { Rectangular } from "../../common/data/rectangle";
import { ChartAreaName } from "./chart-area-name";
import { ChartRegion } from "./chart-region";
import { ChartTile } from "./chart-tile";

export interface ChartObject {
   areaName: ChartAreaName;
   bounds: Rectangular;
   layoutBounds: Rectangular;
   tiles: ChartTile[];
   regions: ChartRegion[];
   containsCustomDcRangeRef?: boolean;
   containsCustomDcMergeRef?: boolean;
}